Making the following configurables an integer so we can use it in list-slicing notations:
  max_length max_emails max_email_size poll_interval

When a float is used in the slicing notation you would see
errors similar to:

>>> 'abc'[1.1:]
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
TypeError: slice indices must be integers or None or have an __index__ method

Thanks Alan for the bug report!

[git-p4: depot-paths = "//guest/lester_cheung/p4review/": change = 10727]

+++ p4review2.py
@@ -240,8 +240,9 @@ def parse_args():
             cfg[key] = true_or_false(cfg.get(key))
             
         for key in 'max_length max_emails max_email_size poll_interval'.split():
-            if key in cfg:
-                cfg[key] = float(cfg.get(key))
+            if key in cfg: 
+                cfg[key] = int(cfg.get(key)) # NOTE: float values cannot be used in list 
+                                             #       slicing notations!
 
         for k in defaults:
             if k in cfg:
